requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
altair==5.5.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
anyio==4.12.0
    # via
    #   httpx
    #   openai
asgiref==3.11.0
    # via django
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
azure-ai-vision-imageanalysis==1.0.0
    # via photo-assistant (pyproject.toml)
azure-cognitiveservices-vision-computervision==0.9.1
    # via photo-assistant (pyproject.toml)
azure-common==1.1.28
    # via azure-cognitiveservices-vision-computervision
azure-core==1.36.0
    # via
    #   photo-assistant (pyproject.toml)
    #   azure-ai-vision-imageanalysis
    #   azure-storage-blob
    #   msrest
azure-storage-blob==12.27.1
    # via photo-assistant (pyproject.toml)
blinker==1.9.0
    # via streamlit
cachetools==6.2.2
    # via
    #   google-auth
    #   streamlit
certifi==2025.11.12
    # via
    #   httpcore
    #   httpx
    #   msrest
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via streamlit
colorama==0.4.6
    # via
    #   click
    #   tqdm
cryptography==46.0.3
    # via azure-storage-blob
distro==1.9.0
    # via openai
django==5.2.8
    # via image
faiss-cpu==1.13.0
    # via photo-assistant (pyproject.toml)
gitdb==4.0.12
    # via gitpython
gitpython==3.1.45
    # via streamlit
google-api-core==2.28.1
    # via google-cloud-vision
google-auth==2.43.0
    # via
    #   google-api-core
    #   google-cloud-vision
google-cloud-vision==3.11.0
    # via photo-assistant (pyproject.toml)
googleapis-common-protos==1.72.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.76.0
    # via
    #   google-api-core
    #   google-cloud-vision
    #   grpcio-status
grpcio-status==1.76.0
    # via google-api-core
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via openai
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
image==1.5.33
    # via photo-assistant (pyproject.toml)
isodate==0.7.2
    # via
    #   azure-ai-vision-imageanalysis
    #   azure-storage-blob
    #   msrest
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
jiter==0.12.0
    # via openai
jsonschema==4.25.1
    # via altair
jsonschema-specifications==2025.9.1
    # via jsonschema
markupsafe==3.0.3
    # via jinja2
msrest==0.7.1
    # via
    #   photo-assistant (pyproject.toml)
    #   azure-cognitiveservices-vision-computervision
narwhals==2.12.0
    # via altair
numpy==2.3.5
    # via
    #   faiss-cpu
    #   pandas
    #   pydeck
    #   streamlit
oauthlib==3.3.1
    # via requests-oauthlib
openai==2.8.1
    # via photo-assistant (pyproject.toml)
packaging==25.0
    # via
    #   altair
    #   faiss-cpu
    #   streamlit
pandas==2.3.3
    # via streamlit
pillow==12.0.0
    # via
    #   photo-assistant (pyproject.toml)
    #   image
    #   streamlit
proto-plus==1.26.1
    # via
    #   google-api-core
    #   google-cloud-vision
protobuf==6.33.1
    # via
    #   google-api-core
    #   google-cloud-vision
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
    #   streamlit
pyarrow==21.0.0
    # via streamlit
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==2.23
    # via cffi
pydantic==2.12.5
    # via openai
pydantic-core==2.41.5
    # via pydantic
pydeck==0.9.1
    # via streamlit
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.2.1
    # via photo-assistant (pyproject.toml)
pytz==2025.2
    # via pandas
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via
    #   photo-assistant (pyproject.toml)
    #   azure-core
    #   google-api-core
    #   msrest
    #   requests-oauthlib
    #   streamlit
requests-oauthlib==2.0.0
    # via msrest
rpds-py==0.29.0
    # via
    #   jsonschema
    #   referencing
rsa==4.9.1
    # via google-auth
six==1.17.0
    # via
    #   image
    #   python-dateutil
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via openai
sqlparse==0.5.4
    # via django
streamlit==1.51.0
    # via photo-assistant (pyproject.toml)
tenacity==9.1.2
    # via streamlit
toml==0.10.2
    # via streamlit
tornado==6.5.2
    # via streamlit
tqdm==4.67.1
    # via openai
typing-extensions==4.15.0
    # via
    #   altair
    #   azure-ai-vision-imageanalysis
    #   azure-core
    #   azure-storage-blob
    #   grpcio
    #   openai
    #   pydantic
    #   pydantic-core
    #   streamlit
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.2
    # via
    #   django
    #   pandas
urllib3==2.5.0
    # via requests
watchdog==6.0.0
    # via streamlit
